new package; move class
